[-] Installer : Decimal replaces Float

+++ install-dev/data/db_structure.sql
@@ -1121,8 +1121,8 @@ CREATE TABLE `PREFIX_order_detail` (
   `product_price` decimal(20,6) NOT NULL default '0.000000',
   `reduction_percent` DECIMAL(10, 2) NOT NULL default '0.00',
   `reduction_amount` DECIMAL(20, 6) NOT NULL default '0.000000',
-  `reduction_amount_tax_incl` FLOAT(20, 6) NOT NULL default '0.000000',
-  `reduction_amount_tax_excl` FLOAT(20, 6) NOT NULL default '0.000000',
+  `reduction_amount_tax_incl` DECIMAL(20, 6) NOT NULL default '0.000000',
+  `reduction_amount_tax_excl` DECIMAL(20, 6) NOT NULL default '0.000000',
   `group_reduction` DECIMAL(10, 2) NOT NULL default '0.000000',
   `product_quantity_discount` decimal(20,6) NOT NULL default '0.000000',
   `product_ean13` varchar(13) default NULL,
@@ -2413,5 +2413,4 @@ CREATE TABLE `PREFIX_module_preference` (
 	`id_shop` int(11) unsigned NOT NULL,
 	PRIMARY KEY (`id_carrier`, `id_tax_rules_group`, `id_shop`)
 ) ENGINE=ENGINE_TYPE  DEFAULT CHARSET=utf8;
-
